pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "docorient"
version = "0.4.0"
description = "Document image orientation detection and correction."
readme = "README.md"
license = "MIT"
requires-python = ">= 3.10"
authors = [{ name = "Lucas Gabriel Vaz" }]
keywords = ["document", "orientation", "rotation", "image", "correction", "onnx"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Typing :: Typed",
]
dependencies = [
    "Pillow >= 10.0",
    "numpy >= 1.24",
    "tqdm >= 4.60",
    "onnxruntime >= 1.16",
]

[project.optional-dependencies]
dev = ["pytest >= 8.0", "pytest-cov", "ruff >= 0.4", "build", "twine"]

[project.scripts]
docorient = "docorient.cli:main"

[project.urls]
Homepage = "https://github.com/lucasleirbag/DocOrient"
Repository = "https://github.com/lucasleirbag/DocOrient"
Issues = "https://github.com/lucasleirbag/DocOrient/issues"
Changelog = "https://github.com/lucasleirbag/DocOrient/blob/main/CHANGELOG.md"

[tool.hatch.build]
exclude = ["/data", "/models", "/scripts"]

[tool.hatch.build.targets.wheel]
packages = ["src/docorient"]

[tool.hatch.build.targets.wheel.force-include]
"src/docorient/models/orientation_detector.onnx" = "docorient/models/orientation_detector.onnx"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "N",
    "UP",
    "B",
    "SIM",
    "RUF",
]
ignore = ["B905"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
